+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## 0.1.1
+- Added `overpaintFactor` parameter for extra paint margin around blobs
+- Added `sharpness` parameter for controlling edge sharpness
+- Normalized threshold to have more effect on alpha cut-off.  
+
 ## 0.1.0
 - Initial release